Aplicativo não abre após seleção de parâmetro para todas as superfícies

+++ app/src/main/java/edu/dhbw/andar/ARObject.java
@@ -162,7 +162,7 @@ public abstract class ARObject {
 		}
 	}
 
-	public abstract float getParameter();
+	public abstract int getParameter();
 
     public abstract void setParameter(float progress);
